+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Check for any possible name collision
 - how to handle dynamic call using function pointer? how to traverse reverse post order?
 - Add a test case for basic block not in topo order. Check whether opt passes will reorder it.
+- Check if PNI graph correctly reflected in summary using a test case.
 
 ## Changelog